+++ substratools/algo.py
@@ -11,6 +11,7 @@ from substratools import exceptions
 from substratools import opener
 from substratools import utils
 from substratools.task_resources import TaskResources
+from substratools.task_resources import StaticInputIdentifiers
 from substratools.workspace import AlgoWorkspace
 
 logger = logging.getLogger(__name__)
@@ -99,18 +100,19 @@ class GenericAlgoWrapper(object):
         # load inputs
         inputs = deepcopy(self._workspace.task_inputs)
 
-        inputs.update({"rank": rank})
+        task_properties = {StaticInputIdentifiers.rank.value: rank}
 
         # load data from opener
         if self._opener_wrapper:
-            X = self._opener_wrapper.get_X(fake_data, n_fake_samples)
-            y = self._opener_wrapper.get_y(fake_data, n_fake_samples)
+            loaded_datasamples = self._opener_wrapper.get_data(fake_data, n_fake_samples)
 
             if fake_data:
                 logger.info("Using fake data with %i fake samples." % int(n_fake_samples))
 
-            inputs.update({"X": X})
-            inputs.update({"y": y})
+            assert (
+                StaticInputIdentifiers.datasamples.value not in inputs.keys()
+            ), f"{StaticInputIdentifiers.datasamples.value} must be an input of kind `datasamples`"
+            inputs.update({StaticInputIdentifiers.datasamples.value: loaded_datasamples})
 
         # load outputs
         outputs = deepcopy(self._workspace.task_outputs)
@@ -122,6 +124,7 @@ class GenericAlgoWrapper(object):
         method(
             inputs=inputs,
             outputs=outputs,
+            task_properties=task_properties,
         )
 
         self._assert_outputs_exists(
@@ -175,6 +178,7 @@ class Algo(GenericAlgo):
         self,
         inputs: dict,
         outputs: dict,
+        task_properties: dict,
     ) -> None:
 
         raise NotImplementedError
@@ -184,6 +188,7 @@ class Algo(GenericAlgo):
         self,
         inputs: dict,
         outputs: dict,
+        task_properties: dict,
     ) -> None:
 
         raise NotImplementedError
@@ -195,6 +200,7 @@ class CompositeAlgo(GenericAlgo):
         self,
         inputs: dict,
         outputs: dict,
+        task_properties: dict,
     ) -> None:
 
         raise NotImplementedError
@@ -204,6 +210,7 @@ class CompositeAlgo(GenericAlgo):
         self,
         inputs: dict,
         outputs: dict,
+        task_properties: dict,
     ) -> None:
 
         raise NotImplementedError
@@ -215,6 +222,7 @@ class AggregateAlgo(GenericAlgo):
         self,
         inputs: dict,
         outputs: dict,
+        task_properties: dict,
     ) -> None:
 
         raise NotImplementedError
@@ -224,6 +232,7 @@ class AggregateAlgo(GenericAlgo):
         self,
         inputs: dict,
         outputs: dict,
+        task_properties: dict,
     ) -> None:
 
         raise NotImplementedError